Update rickybot part 1 combined - post classification and following.ipynb
refactored follow_more_users fn to use pagination from the bluesky api so that feed limit value isn't capped at 100.

The way I have it written the function will break if we reach the end of the feed, need to write some kind of case to handle that, but the feeds I'm using there shouldn't be an end to reach. 
But technically if you set the number of posts to check to like a billion and number of new followers to follow to like a billion it could reach the end and break. 
use responsibly