Update rickybot part 2 combined.ipynb
added in the set() fix finally to make sure that I wouldn't be adding the same users twice in one run. 
A little bit of memory consumption, could have finessed it to share with the query storage but since we're staying in the realm of like 10000 users max it seemed like overkill.
So far seems to be working to make sure I don't have to ctrl+f for user dids in the query when there are duplicates.