Add files via upload
This new file is the sequel to rickybot part 2. So part 4 and 3 are the necessary routines.

In this next phase of the project I am incorporating post caching with our second table. This first run shows that our code works to prevent posts in our current seen set from being reprocessed. The next test will be the same posts but by using the cached version.