Update rickybot part 4 - post caching.ipynb
post cache was successful and observed to stop posts from being processed. Changed the logging just a little so that you can tell which posts are being eliminated by the db cache and which are being eliminated by the currently viewed set, so we can see how necessary the db cache is in the first place.