Add files via upload
next phase of the project. now that post caching worked I'm branching in a new direction - downward. 
for successful posts we want to go through their likers and follow them to get more engagement.

Did a lot of refactoring and cleaned up my code a bit with this new version and now this works.