Update rickybot part 5 - post like diving.ipynb
Penultimate post. Did some odds and ends changes. 

Most important thing was changed ViTFeatureExtractor to ViTImageProcessor to remove deprecation warnings and futureproof the program.

Tested the two classes with the same performance tests that I originally ran FeatureExtractor through (saved from the old colab) and the results were 100% the same so all that needed to be changed was the import and instantiation. 

Final change before moving forward with converting to  microservice will be to get rid of the final huggingface warning about authentications.