Update rickybot_lambda_aggregator.py
small bugfix logging error that was occurring, I set it to skip the warning on friday's compile day when it needed to be saturday. I did like that that logging notification showed up though, so we're gonna keep that in there just to make sure that everything is going according to plan.