Update rickybot micro - DELETE.ipynb
First successful deletion run!

It was a little disconcerting not seeing any print statements so I might add some of those in a second. 

Added follow checks to do a sanity check of the follow count before and after the deletions.

Had a bugfix where my stats code was only being executed if there was a previous run in dynamodb, which wouldn't be the case if there was only one deletion run for the day. Pulled that code outside that if block and now it's always going to execute when we're done.