Update rickybot micro - ADD FOLLOWS.ipynb
some minor tweaks to logging - removed a whitespace line before the end statement because i liked the formatting better and would save a bit of space. Had to change error logging accordingly.

The more important change was we cut matplotlib - I realized I only was using mpl for imread to see the images before, and now that I don't do that anymore I could replace it with the already imported PIL and bytesIO. Tested before making the change and this github update is after the first successful run. 
This was a particularly important change because plt was throwing some kind of weird warning in the lambda runs so now I can hopefully just avoid that completely instead of trying to fix it.